+++ example/lib/special_text/at_text.dart
@@ -2,8 +2,6 @@ import 'package:extended_text_library/extended_text_library.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
-import 'my_special_text_span_builder.dart';
-
 class AtText extends SpecialText {
   static const String flag = "@";
   final int start;
@@ -11,13 +9,16 @@ class AtText extends SpecialText {
   /// whether show background for @somebody
   final bool showAtBackground;
 
-  final BuilderType type;
   AtText(TextStyle textStyle, SpecialTextGestureTapCallback onTap,
-      {this.showAtBackground: false, this.type, this.start})
-      : super(flag, " ", textStyle, onTap: onTap);
+      {this.showAtBackground: false, this.start})
+      : super(
+          flag,
+          " ",
+          textStyle,
+        );
 
   @override
-  TextSpan finishText() {
+  InlineSpan finishText() {
     TextStyle textStyle =
         this.textStyle?.copyWith(color: Colors.blue, fontSize: 16.0);
 
@@ -33,23 +34,19 @@ class AtText extends SpecialText {
             ///caret can move into special text
             deleteAll: true,
             style: textStyle,
-            recognizer: type == BuilderType.extendedText
-                ? (TapGestureRecognizer()
-                  ..onTap = () {
-                    if (onTap != null) onTap(atText);
-                  })
-                : null)
+            recognizer: (TapGestureRecognizer()
+              ..onTap = () {
+                if (onTap != null) onTap(atText);
+              }))
         : SpecialTextSpan(
             text: atText,
             actualText: atText,
             start: start,
             style: textStyle,
-            recognizer: type == BuilderType.extendedText
-                ? (TapGestureRecognizer()
-                  ..onTap = () {
-                    if (onTap != null) onTap(atText);
-                  })
-                : null);
+            recognizer: (TapGestureRecognizer()
+              ..onTap = () {
+                if (onTap != null) onTap(atText);
+              }));
   }
 }