+++ example/lib/special_text/dollar_text.dart
@@ -2,18 +2,15 @@ import 'package:extended_text_library/extended_text_library.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
-import 'my_special_text_span_builder.dart';
-
 class DollarText extends SpecialText {
   static const String flag = "\$";
   final int start;
-  final BuilderType type;
   DollarText(TextStyle textStyle, SpecialTextGestureTapCallback onTap,
-      {this.type, this.start})
+      {this.start})
       : super(flag, flag, textStyle, onTap: onTap);
 
   @override
-  TextSpan finishText() {
+  InlineSpan finishText() {
     final String text = getContent();
 
     return SpecialTextSpan(