+++ lib/extended_text.dart
@@ -4,7 +4,4 @@ export 'src/extended_text.dart';
 export 'src/over_flow_text_span.dart';
 export 'src/extended_render_paragraph.dart';
 export 'src/selection/extended_text_selection_pointer_handler.dart';
-export 'src/selection/selection_controls/extended_text_selection_controls.dart';
-export 'src/selection/selection_controls/material_text_selection_controls.dart';
-export 'src/selection/selection_controls/cupertino_text_selection_controls.dart';
 export 'package:extended_text_library/extended_text_library.dart';