+++ lib/src/extended_render_paragraph.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'dart:ui' as ui show Gradient, Shader, TextBox, PlaceholderAlignment;
+import 'dart:ui' as ui show Gradient, Shader, TextBox;
 import 'dart:math';
 
 import 'extended_text_typedef.dart';
@@ -12,10 +12,8 @@ import 'extended_text_typedef.dart';
 const String _kEllipsis = '\u2026';
 
 /// A render object that displays a paragraph of text
-class ExtendedRenderParagraph extends RenderBox
-    with
-        ContainerRenderObjectMixin<RenderBox, TextParentData>,
-        RenderBoxContainerDefaultsMixin<RenderBox, TextParentData> {
+class ExtendedRenderParagraph extends ExtendedTextRenderBox
+    with ExtendedTextSelectionRenderObject {
   /// Creates a paragraph render object.
   ///
   /// The [text], [textAlign], [textDirection], [overflow], [softWrap], and
@@ -68,12 +66,7 @@ class ExtendedRenderParagraph extends RenderBox
         _selectionColor = selectionColor,
         _selection = selection {
     addAll(children);
-    _extractPlaceholderSpans(text);
-  }
-  @override
-  void setupParentData(RenderBox child) {
-    if (child.parentData is! TextParentData)
-      child.parentData = TextParentData();
+    extractPlaceholderSpans(text);
   }
 
   ///TextSelection
@@ -137,29 +130,17 @@ class ExtendedRenderParagraph extends RenderBox
         return;
       case RenderComparison.paint:
         _textPainter.text = value;
-        _extractPlaceholderSpans(value);
+        extractPlaceholderSpans(value);
         markNeedsPaint();
         markNeedsSemanticsUpdate();
         break;
       case RenderComparison.layout:
         _textPainter.text = value;
         _overflowShader = null;
-        _extractPlaceholderSpans(value);
+        extractPlaceholderSpans(value);
         markNeedsLayout();
         break;
     }
-  }
-
-  List<PlaceholderSpan> _placeholderSpans;
-  void _extractPlaceholderSpans(InlineSpan span) {
-    _placeholderSpans = <PlaceholderSpan>[];
-    span.visitChildren((InlineSpan span) {
-      if (span is PlaceholderSpan) {
-        final PlaceholderSpan placeholderSpan = span;
-        _placeholderSpans.add(placeholderSpan);
-      }
-      return true;
-    });
   }
 
   /// How the text should be aligned horizontally.
@@ -199,6 +180,7 @@ class ExtendedRenderParagraph extends RenderBox
   ///
   /// If [softWrap] is false, [overflow] and [textAlign] may have unexpected
   /// effects.
+  @override
   bool get softWrap => _softWrap;
   bool _softWrap;
   set softWrap(bool value) {
@@ -209,6 +191,7 @@ class ExtendedRenderParagraph extends RenderBox
   }
 
   /// How visual overflow should be handled.
+  @override
   TextOverflow get overflow => _overflow;
   TextOverflow _overflow;
   set overflow(TextOverflow value) {
@@ -286,136 +269,12 @@ class ExtendedRenderParagraph extends RenderBox
   }
 
   @override
-  double computeMinIntrinsicWidth(double height) {
-    if (!_canComputeIntrinsics()) {
-      return 0.0;
-    }
-    _computeChildrenWidthWithMinIntrinsics(height);
-    _layoutText(); // layout with infinite width.
-    return _textPainter.minIntrinsicWidth;
-  }
-
-  @override
-  double computeMaxIntrinsicWidth(double height) {
-    if (!_canComputeIntrinsics()) {
-      return 0.0;
-    }
-    _computeChildrenWidthWithMaxIntrinsics(height);
-    _layoutText(); // layout with infinite width.
-    return _textPainter.maxIntrinsicWidth;
-  }
-
-  double _computeIntrinsicHeight(double width) {
-    if (!_canComputeIntrinsics()) {
-      return 0.0;
-    }
-    _computeChildrenHeightWithMinIntrinsics(width);
-    _layoutText(minWidth: width, maxWidth: width);
-    return _textPainter.height;
-  }
-
-  @override
-  double computeMinIntrinsicHeight(double width) {
-    return _computeIntrinsicHeight(width);
-  }
-
-  @override
-  double computeMaxIntrinsicHeight(double width) {
-    return _computeIntrinsicHeight(width);
-  }
-
-  @override
   double computeDistanceToActualBaseline(TextBaseline baseline) {
     assert(!debugNeedsLayout);
     assert(constraints != null);
     assert(constraints.debugAssertIsValid());
     _layoutTextWithConstraints(constraints);
     return _textPainter.computeDistanceToActualBaseline(baseline);
-  }
-
-  // Intrinsics cannot be calculated without a full layout for
-  // alignments that require the baseline (baseline, aboveBaseline,
-  // belowBaseline).
-  bool _canComputeIntrinsics() {
-    for (PlaceholderSpan span in _placeholderSpans) {
-      switch (span.alignment) {
-        case ui.PlaceholderAlignment.baseline:
-        case ui.PlaceholderAlignment.aboveBaseline:
-        case ui.PlaceholderAlignment.belowBaseline:
-          {
-            assert(
-                RenderObject.debugCheckingIntrinsics,
-                'Intrinsics are not available for PlaceholderAlignment.baseline, '
-                'PlaceholderAlignment.aboveBaseline, or PlaceholderAlignment.belowBaseline,');
-            return false;
-          }
-        case ui.PlaceholderAlignment.top:
-        case ui.PlaceholderAlignment.middle:
-        case ui.PlaceholderAlignment.bottom:
-          {
-            continue;
-          }
-      }
-    }
-    return true;
-  }
-
-  void _computeChildrenWidthWithMaxIntrinsics(double height) {
-    RenderBox child = firstChild;
-    final List<PlaceholderDimensions> placeholderDimensions =
-        List<PlaceholderDimensions>(childCount);
-    int childIndex = 0;
-    while (child != null) {
-      // Height and baseline is irrelevant as all text will be laid
-      // out in a single line.
-      placeholderDimensions[childIndex] = PlaceholderDimensions(
-        size: Size(child.getMaxIntrinsicWidth(height), height),
-        alignment: _placeholderSpans[childIndex].alignment,
-        baseline: _placeholderSpans[childIndex].baseline,
-      );
-      child = childAfter(child);
-      childIndex += 1;
-    }
-    _textPainter.setPlaceholderDimensions(placeholderDimensions);
-  }
-
-  void _computeChildrenWidthWithMinIntrinsics(double height) {
-    RenderBox child = firstChild;
-    final List<PlaceholderDimensions> placeholderDimensions =
-        List<PlaceholderDimensions>(childCount);
-    int childIndex = 0;
-    while (child != null) {
-      final double intrinsicWidth = child.getMinIntrinsicWidth(height);
-      final double intrinsicHeight =
-          child.getMinIntrinsicHeight(intrinsicWidth);
-      placeholderDimensions[childIndex] = PlaceholderDimensions(
-        size: Size(intrinsicWidth, intrinsicHeight),
-        alignment: _placeholderSpans[childIndex].alignment,
-        baseline: _placeholderSpans[childIndex].baseline,
-      );
-      child = childAfter(child);
-      childIndex += 1;
-    }
-    _textPainter.setPlaceholderDimensions(placeholderDimensions);
-  }
-
-  void _computeChildrenHeightWithMinIntrinsics(double width) {
-    RenderBox child = firstChild;
-    final List<PlaceholderDimensions> placeholderDimensions =
-        List<PlaceholderDimensions>(childCount);
-    int childIndex = 0;
-    while (child != null) {
-      final double intrinsicHeight = child.getMinIntrinsicHeight(width);
-      final double intrinsicWidth = child.getMinIntrinsicWidth(intrinsicHeight);
-      placeholderDimensions[childIndex] = PlaceholderDimensions(
-        size: Size(intrinsicWidth, intrinsicHeight),
-        alignment: _placeholderSpans[childIndex].alignment,
-        baseline: _placeholderSpans[childIndex].baseline,
-      );
-      child = childAfter(child);
-      childIndex += 1;
-    }
-    _textPainter.setPlaceholderDimensions(placeholderDimensions);
   }
 
   @override
@@ -496,88 +355,15 @@ class ExtendedRenderParagraph extends RenderBox
   @visibleForTesting
   bool get debugHasOverflowShader => _overflowShader != null;
 
-  void _layoutText({double minWidth = 0.0, double maxWidth = double.infinity}) {
-    final bool widthMatters = softWrap || overflow == TextOverflow.ellipsis;
-    _textPainter.layout(
-        minWidth: minWidth,
-        maxWidth: widthMatters ? maxWidth : double.infinity);
-  }
-
   void _layoutTextWithConstraints(BoxConstraints constraints) {
-    _layoutText(minWidth: constraints.minWidth, maxWidth: constraints.maxWidth);
-  }
-
-  // Layout the child inline widgets. We then pass the dimensions of the
-  // children to _textPainter so that appropriate placeholders can be inserted
-  // into the LibTxt layout. This does not do anything if no inline widgets were
-  // specified.
-  void _layoutChildren(BoxConstraints constraints) {
-    if (childCount == 0) {
-      return;
-    }
-    RenderBox child = firstChild;
-    final List<PlaceholderDimensions> placeholderDimensions =
-        List<PlaceholderDimensions>(childCount);
-    int childIndex = 0;
-    while (child != null) {
-      // Only constrain the width to the maximum width of the paragraph.
-      // Leave height unconstrained, which will overflow if expanded past.
-      child.layout(
-          BoxConstraints(
-            maxWidth: constraints.maxWidth,
-          ),
-          parentUsesSize: true);
-      double baselineOffset;
-      switch (_placeholderSpans[childIndex].alignment) {
-        case ui.PlaceholderAlignment.baseline:
-          {
-            baselineOffset = child
-                .getDistanceToBaseline(_placeholderSpans[childIndex].baseline);
-            break;
-          }
-        default:
-          {
-            baselineOffset = null;
-            break;
-          }
-      }
-      placeholderDimensions[childIndex] = PlaceholderDimensions(
-        size: child.size,
-        alignment: _placeholderSpans[childIndex].alignment,
-        baseline: _placeholderSpans[childIndex].baseline,
-        baselineOffset: baselineOffset,
-      );
-      child = childAfter(child);
-      childIndex += 1;
-    }
-    _textPainter.setPlaceholderDimensions(placeholderDimensions);
-  }
-
-  // Iterate through the laid-out children and set the parentData offsets based
-  // off of the placeholders inserted for each child.
-  void _setParentData() {
-    RenderBox child = firstChild;
-    int childIndex = 0;
-
-    ///maybe overflow
-    while (child != null &&
-        childIndex < _textPainter.inlinePlaceholderBoxes.length) {
-      final TextParentData textParentData = child.parentData;
-
-      textParentData.offset = Offset(
-          _textPainter.inlinePlaceholderBoxes[childIndex].left,
-          _textPainter.inlinePlaceholderBoxes[childIndex].top);
-      textParentData.scale = _textPainter.inlinePlaceholderScales[childIndex];
-      child = childAfter(child);
-      childIndex += 1;
-    }
+    layoutText(minWidth: constraints.minWidth, maxWidth: constraints.maxWidth);
   }
 
   @override
   void performLayout() {
-    _layoutChildren(constraints);
+    layoutChildren(constraints);
     _layoutTextWithConstraints(constraints);
-    _setParentData();
+    setParentData();
 
     // We grab _textPainter.size and _textPainter.didExceedMaxLines here because
     // assigning to `size` will trigger us to validate our intrinsic sizes,
@@ -651,11 +437,8 @@ class ExtendedRenderParagraph extends RenderBox
     }
   }
 
-  Offset _initialOffset;
-  Offset get _effectiveOffset => (_initialOffset ?? Offset.zero);
   @override
   void paint(PaintingContext context, Offset offset) {
-    _initialOffset = offset;
     // Ideally we could compute the min/max intrinsic width/height with a
     // non-destructive operation. However, currently, computing these values
     // will destroy state inside the painter. If that happens, we need to
@@ -704,31 +487,8 @@ class ExtendedRenderParagraph extends RenderBox
     }
     _textPainter.paint(context.canvas, offset);
 
-    RenderBox child = firstChild;
-    int childIndex = 0;
+    paintWidgets(context, offset);
 
-    ///maybe overflow
-    while (child != null &&
-        childIndex < _textPainter.inlinePlaceholderBoxes.length) {
-      //assert(childIndex < _textPainter.inlinePlaceholderBoxes.length);
-
-      final TextParentData textParentData = child.parentData;
-
-      final double scale = textParentData.scale;
-      context.pushTransform(
-        needsCompositing,
-        offset + textParentData.offset,
-        Matrix4.diagonal3Values(scale, scale, scale),
-        (PaintingContext context, Offset offset) {
-          context.paintChild(
-            child,
-            offset,
-          );
-        },
-      );
-      child = childAfter(child);
-      childIndex += 1;
-    }
     if (_needsClipping) {
       if (_overflowShader != null) {
         context.canvas.translate(offset.dx, offset.dy);
@@ -1206,11 +966,9 @@ class ExtendedRenderParagraph extends RenderBox
       selection = TextSelection(baseOffset: word.start, extentOffset: word.end);
     }
     return handleSpecialText
-        ? convertTextPainterSelectionToTextInputSelection(text, selection)
+        ? convertTextPainterSelectionToTextInputSelection(text, selection,
+        selectWord: true)
         : selection;
-//    if (position.offset >= word.end)
-//      return TextSelection.fromPosition(position);
-//    return TextSelection(baseOffset: word.start, extentOffset: word.end);
   }
 
   /// Move the selection to the beginning or end of a word.
@@ -1358,8 +1116,7 @@ class ExtendedRenderParagraph extends RenderBox
   ///
   ///  * [getLocalRectForCaret], which is the equivalent but for
   ///    a [TextPosition] rather than a [TextSelection].
-  List<TextSelectionPoint> getEndpointsForSelection(TextSelection selection,
-      {bool toolbar: false}) {
+  List<TextSelectionPoint> getEndpointsForSelection(TextSelection selection) {
     assert(constraints != null);
 //    if (temp.isCollapsed && toolbar) {
 //      // TODO(mpcomplete): This doesn't work well at an RTL/LTR boundary.
@@ -1460,12 +1217,14 @@ class ExtendedRenderParagraph extends RenderBox
 
     final Rect visibleRegion = Offset.zero & size;
 
-    final Offset startOffset = _getCaretOffset(
-        TextPosition(
-          offset: selection.start,
-          affinity: selection.affinity,
-        ),
-        effectiveOffset: effectiveOffset);
+    final Offset startOffset = getCaretOffset(
+      TextPosition(
+        offset: selection.start,
+        affinity: selection.affinity,
+      ),
+      effectiveOffset: effectiveOffset,
+      handleSpecialText: handleSpecialText,
+    );
 
     // TODO(justinmc): https://github.com/flutter/flutter/issues/31495
     // Check if the selection is visible with an approximation because a
@@ -1479,9 +1238,11 @@ class ExtendedRenderParagraph extends RenderBox
         .inflate(visibleRegionSlop)
         .contains(startOffset + effectiveOffset);
 
-    final Offset endOffset = _getCaretOffset(
-        TextPosition(offset: selection.end, affinity: selection.affinity),
-        effectiveOffset: effectiveOffset);
+    final Offset endOffset = getCaretOffset(
+      TextPosition(offset: selection.end, affinity: selection.affinity),
+      effectiveOffset: effectiveOffset,
+      handleSpecialText: handleSpecialText,
+    );
 
     _selectionEndInViewport.value = visibleRegion
         .inflate(visibleRegionSlop)
@@ -1493,38 +1254,19 @@ class ExtendedRenderParagraph extends RenderBox
     return visibleRegion.contains(globalToLocal(position));
   }
 
-  Offset _getCaretOffset(TextPosition textPosition,
-      {ValueChanged<double> caretHeightCallBack, Offset effectiveOffset}) {
-    effectiveOffset ??= this._effectiveOffset;
-
-    ///zmt
-    if (handleSpecialText) {
-      ///if first index, check by first span
-      var offset = textPosition.offset;
-      if (offset == 0) {
-        offset = 1;
-      }
-
-      ///last or has ExtendedWidgetSpan
-
-      var boxs = _textPainter.getBoxesForSelection(TextSelection(
-          baseOffset: offset - 1,
-          extentOffset: offset,
-          affinity: textPosition.affinity));
-      if (boxs.length > 0) {
-        var rect = boxs.toList().last.toRect();
-        caretHeightCallBack?.call(rect.height);
-        if (textPosition.offset == 0) {
-          return rect.topLeft + effectiveOffset;
-        } else {
-          return rect.topRight + effectiveOffset;
-        }
-      }
-    }
-
-    final Offset caretOffset =
-        _textPainter.getOffsetForCaret(textPosition, Rect.zero) +
-            effectiveOffset;
-    return caretOffset;
+  @override
+  Size getSize() {
+    return this.size;
   }
+
+  @override
+  bool get isAttached => attached;
+
+  @override
+  Offset getlocalToGlobal(Offset point, {RenderObject ancestor}) {
+    return localToGlobal(point, ancestor: ancestor);
+  }
+
+  @override
+  TextPainter get textPainter => _textPainter;
 }