+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.6.4]
+
+* Improve codes about selection
+* Select all SpecialTextSpan which deleteAll is true when double tap or long tap
+
 ## [0.6.3]
 
 * Fix issue [ImageSpan is not TextSpan](https://github.com/fluttercandies/extended_text/issues/24)